cast-service/pyproject.toml
[tool.poetry]
name = "cast-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.8.1"
python-dotenv = "^1.0.0"
asyncpg = "0.20.1"
databases = {version = "0.2.6", extras = ["postgresql"]}
fastapi = {version = "0.48.0", extras = ['all']}
sqlalchemy = "1.3.13"
uvicorn = "0.11.2"

[tool.poetry.scripts]
start = "app.main:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
